dom/media/webaudio/test/test_analyserNodeWithGain.html

<!DOCTYPE html>
<title>Test effect of AnalyserNode on GainNode output</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
  // fftSize <= 128 so that the time domain data is full of input after
  // notification that the first block has been processed.
  const fftSize = 32;

  var context = new AudioContext();

  var analyser1 = context.createAnalyser();
  analyser1.fftSize = fftSize;
  var analyser2 = context.createAnalyser();
  analyser2.fftSize = fftSize;

  var gain = context.createGain();
  gain.gain.value = 2.0;
  gain.connect(analyser1);
  gain.connect(analyser2);

  // Create a DC input to make getFloatTimeDomainData() output consistent at
  // any time.
  var buffer = context.createBuffer(1, 1, context.sampleRate);
  buffer.getChannelData(0)[0] = 1.0 / gain.gain.value;
  var source = context.createBufferSource();
  source.buffer = buffer;
  source.loop = true;
  source.connect(gain);
  source.start();

  // Waiting for an ended event ensures that the AnalyserNode has received the
  // signal.
  var timer = context.createBufferSource();
  timer.buffer = buffer;
  timer.onended = t.step_func_done(function() {
    var data = new Float32Array(1);
    analyser1.getFloatTimeDomainData(data);
    assert_equals(data[0], 1.0, "analyser1 time domain data");
    analyser2.getFloatTimeDomainData(data);
    assert_equals(data[0], 1.0, "analyser2 time domain data");
  });

  timer.start()
});
</script>
